+++ src/mahoji/commands/minion.ts
@@ -2,8 +2,6 @@ import { ApplicationCommandOptionType, CommandRunOptions } from 'mahoji';
 
 import { client } from '../..';
 import { diaries } from '../../lib/diaries';
-import { KourendFavours } from '../../lib/minions/data/kourendFavour';
-import Potions from '../../lib/minions/data/potions';
 import getOSItem from '../../lib/util/getOSItem';
 import { minionStatsEmbed } from '../../lib/util/minionStatsEmbed';
 import BankImageTask from '../../tasks/bankImage';
@@ -12,15 +10,11 @@ import {
 	claimAchievementDiaryCommand
 } from '../lib/abstracted_commands/achievementDiaryCommand';
 import { bankBgCommand } from '../lib/abstracted_commands/bankBgCommand';
-import { collectables, collectCommand } from '../lib/abstracted_commands/collectCommand';
 import { crackerCommand } from '../lib/abstracted_commands/crackerCommand';
-import { decantCommand } from '../lib/abstracted_commands/decantCommand';
-import { favourCommand } from '../lib/abstracted_commands/favourCommand';
 import { Lampables, lampCommand } from '../lib/abstracted_commands/lampCommand';
-import { questCommand } from '../lib/abstracted_commands/questCommand';
 import { skillOption } from '../lib/mahojiCommandOptions';
 import { OSBMahojiCommand } from '../lib/util';
-import { MahojiUserOption, mahojiUsersSettingsFetch } from '../mahojiSettings';
+import { MahojiUserOption } from '../mahojiSettings';
 
 export const minionCommand: OSBMahojiCommand = {
 	name: 'minion',
@@ -90,75 +84,6 @@ export const minionCommand: OSBMahojiCommand = {
 		},
 		{
 			type: ApplicationCommandOptionType.Subcommand,
-			name: 'quest',
-			description: 'Send your minion to do quests.'
-		},
-		{
-			type: ApplicationCommandOptionType.Subcommand,
-			name: 'favour',
-			description: 'Allows you to get Kourend Favour.',
-			options: [
-				{
-					type: ApplicationCommandOptionType.String,
-					name: 'name',
-					description: 'The name of the Kourend House.',
-					choices: KourendFavours.map(i => ({ name: i.name, value: i.name })),
-					required: false
-				},
-				{
-					type: ApplicationCommandOptionType.Boolean,
-					name: 'no_stams',
-					description: "Enable if you don't want to use stamina potions when getting favour.",
-					required: false
-				}
-			]
-		},
-		{
-			type: ApplicationCommandOptionType.Subcommand,
-			name: 'decant',
-			description: 'Allows you to decant potions into different dosages.',
-			options: [
-				{
-					type: ApplicationCommandOptionType.String,
-					name: 'potion_name',
-					description: 'The name of the bank background you want.',
-					autocomplete: async (value: string) => {
-						return Potions.filter(p =>
-							!value ? true : p.name.toLowerCase().includes(value.toLowerCase())
-						).map(p => ({ name: p.name, value: p.name }));
-					},
-					required: true
-				},
-				{
-					type: ApplicationCommandOptionType.Number,
-					name: 'dose',
-					description: 'The dosage to decant them too. (default 4)',
-					required: false,
-					min_value: 1,
-					max_value: 4
-				}
-			]
-		},
-		{
-			type: ApplicationCommandOptionType.Subcommand,
-			name: 'collect',
-			description: 'Sends your minion to collect items.',
-			options: [
-				{
-					type: ApplicationCommandOptionType.String,
-					name: 'item',
-					description: 'The item you want to collect.',
-					autocomplete: async (value: string) => {
-						return collectables
-							.filter(p => (!value ? true : p.item.name.toLowerCase().includes(value.toLowerCase())))
-							.map(p => ({ name: p.item.name, value: p.item.name }));
-					},
-					required: true
-				}
-			]
-		},
-		{
-			type: ApplicationCommandOptionType.Subcommand,
 			name: 'lamp',
 			description: 'Use lamps to claim XP.',
 			options: [
@@ -195,21 +120,15 @@ export const minionCommand: OSBMahojiCommand = {
 	run: async ({
 		userID,
 		options,
-		channelID,
 		interaction
 	}: CommandRunOptions<{
 		stats?: {};
 		achievementdiary?: { diary?: string; claim?: boolean };
 		bankbg?: { name?: string };
 		cracker?: { user: MahojiUserOption };
-		quest?: {};
-		favour?: { name?: string; no_stams?: boolean };
-		decant?: { potion_name: string; dose?: number };
-		collect?: { item: string };
 		lamp?: { item: string; quantity?: number; skill: string };
 	}>) => {
 		const user = await client.fetchUser(userID.toString());
-		const mahojiUser = await mahojiUsersSettingsFetch(userID);
 
 		if (options.stats) {
 			return { embeds: [await minionStatsEmbed(user)] };
@@ -228,18 +147,6 @@ export const minionCommand: OSBMahojiCommand = {
 		if (options.cracker) {
 			const otherUser = await client.fetchUser(options.cracker.user.user.id);
 			return crackerCommand({ owner: user, otherPerson: otherUser, interaction });
-		}
-		if (options.quest) {
-			return questCommand(user, channelID);
-		}
-		if (options.favour) {
-			return favourCommand(user, mahojiUser, options.favour.name, channelID, options.favour.no_stams);
-		}
-		if (options.decant) {
-			return decantCommand(user, options.decant.potion_name, options.decant.dose);
-		}
-		if (options.collect) {
-			return collectCommand(mahojiUser, user, channelID, options.collect.item);
 		}
 
 		if (options.lamp) {
