+++ src/tasks/minions/minigames/chompyHuntActivity.ts
@@ -2,11 +2,11 @@ import { roll } from 'e';
 import { Task } from 'klasa';
 import { Bank } from 'oldschooljs';
 
-import { chompyHats } from '../../../commands/Minion/chompyhunt';
 import { userhasDiaryTier, WesternProv } from '../../../lib/diaries';
 import { getMinigameEntity, incrementMinigameScore } from '../../../lib/settings/settings';
 import { MinigameActivityTaskOptions } from '../../../lib/types/minions';
 import { handleTripFinish } from '../../../lib/util/handleTripFinish';
+import { chompyHats } from '../../../mahoji/lib/abstracted_commands/chompyHuntCommand';
 
 export default class extends Task {
 	async run(data: MinigameActivityTaskOptions) {
